Changed Exception classes to public classes as they are on the interface of Vdm & Nmea classes and currently not visible for code in other packages then com.aisparser
