+++ java/com/aisparser/ChecksumFailedException.java
@@ -1,0 +1,11 @@
+package com.aisparser;
+
+public class ChecksumFailedException extends Exception
+{
+    public ChecksumFailedException() {}
+      public ChecksumFailedException( String str )
+        {
+              super(str);
+                }
+}
+
