+++ java/com/aisparser/IllegalNMEACharacterException.java
@@ -1,0 +1,11 @@
+package com.aisparser;
+
+public class IllegalNMEACharacterException extends Exception
+{
+    public IllegalNMEACharacterException() {}
+    public IllegalNMEACharacterException( String str )
+    {
+        super(str);
+    }
+}
+
