+++ java/com/aisparser/StartNotFoundException.java
@@ -1,0 +1,11 @@
+package com.aisparser;
+
+public class StartNotFoundException extends Exception
+{
+    public StartNotFoundException() {}
+    public StartNotFoundException( String str )
+    {
+       super(str);
+    }
+}
+
