+++ java/com/aisparser/VDMSentenceException.java
@@ -1,0 +1,11 @@
+package com.aisparser;
+
+public class VDMSentenceException extends Exception
+{
+    public VDMSentenceException() {}
+    public VDMSentenceException( String str )
+    {
+         super(str);
+    }
+}
+
